requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==1.10.8
SQLAlchemy==2.0.15
PyJWT==2.7.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
